lambda/py/requirements.txt
ask-sdk-core==1.13.0
ask-sdk-model==1.23.0
ask-sdk-runtime==1.13.0
ask-sdk-s3-persistence-adapter==1.0.0
boto3==1.13.1
botocore==1.16.1
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
idna==2.9
jmespath==0.9.5
python-dateutil==2.8.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
